finished writing the confusion matrix which is also used in calculating the accuracy of the classification

+++ src/training_module.cpp
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-vector<vector<vector<int>>> module;
+vector<vector<vector<double>>> module;
 vector<double> prior;
 
 ifstream ins;
